ctfs/RACTF/2021/web/Secret_Store/pyproject.toml
[tool.poetry]
name = "secretstore"
version = "0.1.0"
description = ""
authors = ["David Cooke <me@dave.lc>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.12.4"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
